Entities now have an internal communication system between components, based on events, behaviours and actions.
Events:
Events are messages that contain information about things such as the entity being stunned by an ability,
damaged from stepping on a spike, or something as simple as just having moved.

Actions:
Actions are messages that contain information about an action the entity is to perform, which are created by behaviours.
Actions can be messages such as "Move right 10 pixels" or "Take 10 damage", which are created by behaviour objects, and executed
by entities on every "update()".

Behaviours:
Behaviours are objects that entities possess, which during the entities "preUpdate()" are updated to create actions.
Behaviours have a lifetime that is decreased for every update. When a behaviour is updated it
returns a list of actions to the entity, that it stores to be executed when the entity itself is updated.
An example of a behaviour is a "Dash" behaviour, where the lifetime might be as short as one second. When the dash behaviour is updated
it returns an action to move a set amount of pixels in a direction, and the lifetime is decreased by the amount of time that frame (Actually the previous frame) took to execute.
As a result, the Dash behaviour makes the character move quickly in a direction, over a short period of time.

Behaviour maps:
Behaviour maps is the way an entity creates behaviours. On every "preUpdate()", the map is executed on the set of events the entity has.
The map contains connections from types of events, to types of behaviours.
An example are the connections to the Dash Behaviour, where one connection is that there must exist a "pressedDirectionKey" event, a "pressedDashKey" event,
that the event "stunned" cant exist. If these connections are satisfied, the behaviour "Dash" is created and initialized with the prementioned events themselves,
where in this example, the behaviour retrieves the "angle" parameter of the pressedDirectionKey event to find out which direction to move.

Also started working on a config system. Where config files can be read on startup to modify values in the game, such as keybindings and graphics settings.

+++ AbyssEngine/graphics/buffers/VertexArray.cpp
@@ -8,7 +8,7 @@ VertexArray::VertexArray()
 
 VertexArray::~VertexArray()
 {
-	for (int i = 0; i < buffers.size(); i++) 
+	for (size_t i = 0; i < buffers.size(); i++) 
 	{
 		delete buffers[i];
 	}
